+++ R/cre.R
@@ -123,7 +123,7 @@ cre <- function(y, z, X, ratio_dis, ite_method_dis, include_ps_dis = NA,
   }
 
   ite_method_dis <- tolower(ite_method_dis)
-  if (!(ite_method_dis %in% c("ipw", "sipw", "sipw", "or", "bart", "xbart",
+  if (!(ite_method_dis %in% c("ipw", "sipw", "aipw", "or", "bart", "xbart",
                               "bcf", "xbcf", "cf", "poisson"))) {
     stop(paste("Invalid ITE method for Discovery Subsample. Please choose ",
                "from the following:\n","'ipw', 'sipw', 'aipw', or', 'bart', ",
@@ -215,7 +215,7 @@ cre <- function(y, z, X, ratio_dis, ite_method_dis, include_ps_dis = NA,
   }
 
   # Check for correct offset input
-  if (ite_method_dis == "poisson" | ite_method_inf == "poisson" | cate_method == "poisson") {
+  if ((ite_method_dis == "poisson") | (ite_method_inf == "poisson") | (cate_method == "poisson")) {
     if (include_offset == TRUE) {
       if (is.na(offset_name)) {
         stop(paste("Invalid offset_name input. Please specify an offset_name ",