+++ R/estimate_ite.R
@@ -93,9 +93,7 @@ estimate_ite <- function(y, z, X, ite_method, include_ps, ps_method, or_method,
     ite <- ite_results[[1]]
     sd_ite <- ite_results[[2]]
   } else if (ite_method == "poisson") {
-    ite_results <- estimate_ite_poisson(y, z, X, X_names,
-                                        include_offset, offset_name)
-    ite <- ite_results[[1]]
+    ite <- estimate_ite_poisson(y, z, X, X_names, include_offset, offset_name)
     sd_ite <- NA
   } else {
     stop(paste("Invalid ITE method. Please choose from the following:\n",